refactor: remove IMiniFL interface and DeltaNeutralVault02
