--- contracts/8.10/protocol/interfaces/IMiniFL.sol
@@ -1,18 +1,0 @@
-// SPDX-License-Identifier: BUSL
-/**
-  ∩~~~~∩ 
-  ξ ･×･ ξ 
-  ξ　~　ξ 
-  ξ　　 ξ 
-  ξ　　 “~～~～〇 
-  ξ　　　　　　 ξ 
-  ξ ξ ξ~～~ξ ξ ξ 
-　 ξ_ξξ_ξ　ξ_ξξ_ξ
-Alpaca Fin Corporation
-*/
-
-pragma solidity 0.8.10;
-
-interface IMiniFL {
-  function harvest(uint256 _pid) external;
-}
